backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
pydantic[email]
google-genai
PyMuPDF
chromadb
scikit-learn
PyYAML
